# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TranslationInput {
  /** English text to translate */
  text: string;
  /**
     * Optional glossary ID to use for consistent term translation
     * @nullable
     */
  glossaryId?: number | null;
  /**
     * Optional novel title for context
     * @nullable
     */
  novelTitle?: string | null;
}

export type TranslationResultAppliedTermsItem = {
  english?: string;
  burmese?: string;
};

export interface TranslationResult {
  translatedText: string;
  originalText: string;
  appliedTerms?: TranslationResultAppliedTermsItem[];
}

export interface TranslationHistory {
  id: number;
  originalText: string;
  translatedText: string;
  /** @nullable */
  glossaryId?: number | null;
  /** @nullable */
  novelTitle?: string | null;
  createdAt: string;
}

export interface Glossary {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  novelTitle?: string | null;
  termCount?: number;
  createdAt: string;
}

export interface GlossaryInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  novelTitle?: string | null;
}

export interface GlossaryUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  novelTitle?: string | null;
}

export interface Term {
  id: number;
  glossaryId: number;
  english: string;
  burmese: string;
  /**
     * e.g. character, place, technique, rank
     * @nullable
     */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface GlossaryExport {
  glossary: Glossary;
  terms: Term[];
}

export interface TermInput {
  english: string;
  burmese: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface GlossaryImport {
  glossary: GlossaryInput;
  terms: TermInput[];
}

export interface TermUpdate {
  english?: string;
  burmese?: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
}