# lib/api-zod/src/generated/types/glossary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Glossary {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  novelTitle?: string | null;
  termCount?: number;
  createdAt: string;
}