# lib/api-zod/src/generated/types/glossaryExport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Glossary } from './glossary';
import type { Term } from './term';

export interface GlossaryExport {
  glossary: Glossary;
  terms: Term[];
}