# lib/api-zod/src/generated/types/glossaryImport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GlossaryInput } from './glossaryInput';
import type { TermInput } from './termInput';

export interface GlossaryImport {
  glossary: GlossaryInput;
  terms: TermInput[];
}