# lib/api-zod/src/generated/types/glossaryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GlossaryInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  novelTitle?: string | null;
}