# lib/api-zod/src/generated/types/glossaryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GlossaryUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  novelTitle?: string | null;
}