# lib/api-zod/src/generated/types/term.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Term {
  id: number;
  glossaryId: number;
  english: string;
  burmese: string;
  /**
     * e.g. character, place, technique, rank
     * @nullable
     */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
}