# lib/api-zod/src/generated/types/termInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TermInput {
  english: string;
  burmese: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  notes?: string | null;
}