# lib/api-zod/src/generated/types/translationHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslationHistory {
  id: number;
  originalText: string;
  translatedText: string;
  /** @nullable */
  glossaryId?: number | null;
  /** @nullable */
  novelTitle?: string | null;
  createdAt: string;
}