# lib/api-zod/src/generated/types/translationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslationInput {
  /** English text to translate */
  text: string;
  /**
     * Optional glossary ID to use for consistent term translation
     * @nullable
     */
  glossaryId?: number | null;
  /**
     * Optional novel title for context
     * @nullable
     */
  novelTitle?: string | null;
}