# lib/api-zod/src/generated/types/translationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TranslationResultAppliedTermsItem } from './translationResultAppliedTermsItem';

export interface TranslationResult {
  translatedText: string;
  originalText: string;
  appliedTerms?: TranslationResultAppliedTermsItem[];
}